[AIP-21] Move operators.hive_stats_operator providers.apache.hive.operators.hive_stats
