+++ scripts/ci/pylint_todo.txt
@@ -147,7 +147,7 @@
 ./airflow/operators/email_operator.py
 ./airflow/operators/generic_transfer.py
 ./airflow/providers/apache/hive/operators/hive.py
-./airflow/operators/hive_stats_operator.py
+./airflow/providers/apache/hive/operators/hive_stats.py
 ./airflow/operators/http_operator.py
 ./airflow/operators/jdbc_operator.py
 ./airflow/operators/latest_only_operator.py
